Separate map_gfx_utils components
- Place the map_gfx_utils components in separate code files
- Fix ODR issue
